+++ Module_4/true_math.py
@@ -2,6 +2,6 @@ from math import inf
 
 
 def divide(first, second):
-    if int(second) == 0:
+    if float(second) == 0.0:
         return inf
     return float(first) / float(second)
